docs added + readme added + submodules developed
