+++ pubkey_encrypt.info.yml
@@ -1,5 +1,5 @@
 name: Pubkey Encrypt
-description: Adds Credentials-based Public Key Encryption support.
+description: Provides support for encrypting data with users login-credentials.
 core: 8.x
 type: module
 package: Encryption
