+++ pubkey_encrypt_openssl/pubkey_encrypt_openssl.info.yml
@@ -1,0 +1,5 @@
+name: Pubkey Encrypt OpenSSL
+description: Provides an OpenSSL-based Asymmetric Keys Generator to be used with Pubkey Encrypt.
+core: 8.x
+type: module
+package: Encryption
