+++ pubkey_encrypt_openssl/pubkey_encrypt_openssl.install
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * @file
+ * Install hooks for the pubkey_encrypt_openssl module.
+ */
+
+/**
+ * Implements hook_requirements().
+ */
+function pubkey_encrypt_openssl_requirements($phase) {
+  $requirements = array();
+
+  $requirements['pubkey_encrypt_openssl'] = array(
+    'title' => t('Pubkey Encrypt OpenSSL'),
+  );
+
+  if (extension_loaded('openssl')) {
+    $requirements['pubkey_encrypt_openssl']['severity'] = REQUIREMENT_OK;
+  }
+  else {
+    $requirements['pubkey_encrypt_openssl']['severity'] = REQUIREMENT_ERROR;
+    $requirements['pubkey_encrypt_openssl']['description'] = t('Please install and enable the OpenSSL PHP extension.');
+  }
+
+  return $requirements;
+}
