+++ pubkey_encrypt_password/README.md
@@ -1,0 +1,4 @@
+pubkey_encrypt_password
+=======================
+
+Provides a user passwords-based Login Credentials Provider to be used with Pubkey Encrypt.
