+++ pubkey_encrypt_password/pubkey_encrypt_password.info.yml
@@ -1,0 +1,5 @@
+name: Pubkey Encrypt Password
+description: Provides a user passwords-based Login Credentials Provider to be used with Pubkey Encrypt.
+core: 8.x
+type: module
+package: Encryption
