+++ pubkey_encrypt_phpseclib/pubkey_encrypt_phpseclib.info.yml
@@ -3,5 +3,4 @@ description: Provides a PHPSecLib-based Asymmetric Keys Generator to be used wit
 core: 8.x
 type: module
 package: Encryption
-dependencies:
-  - pubkey_encrypt
+
